move helper methods to utils class to avoid too early static init of deperecation logger
