+++ server/src/main/java/org/elasticsearch/common/time/DateFormatters.java
@@ -19,16 +19,12 @@
 
 package org.elasticsearch.common.time;
 
-import org.apache.logging.log4j.LogManager;
 import org.elasticsearch.common.Strings;
-import org.elasticsearch.common.logging.DeprecationLogger;
-import org.joda.time.DateTimeZone;
 
 import java.time.DateTimeException;
 import java.time.DayOfWeek;
 import java.time.Instant;
 import java.time.LocalDate;
-import java.time.ZoneId;
 import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
@@ -40,10 +36,7 @@ import java.time.temporal.IsoFields;
 import java.time.temporal.TemporalAccessor;
 import java.time.temporal.TemporalAdjusters;
 import java.time.temporal.WeekFields;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.Locale;
-import java.util.Map;
 
 import static java.time.temporal.ChronoField.DAY_OF_MONTH;
 import static java.time.temporal.ChronoField.DAY_OF_WEEK;
@@ -1574,43 +1567,5 @@ public class DateFormatters {
         }
 
         return result;
-    }
-
-    public static DateTimeZone zoneIdToDateTimeZone(ZoneId zoneId) {
-        if (zoneId == null) {
-            return null;
-        }
-        if (zoneId instanceof ZoneOffset) {
-            // the id for zoneoffset is not ISO compatible, so cannot be read by ZoneId.of
-            return DateTimeZone.forOffsetMillis(((ZoneOffset)zoneId).getTotalSeconds() * 1000);
-        }
-        return DateTimeZone.forID(zoneId.getId());
-    }
-
-    private static final DeprecationLogger DEPRECATION_LOGGER = new DeprecationLogger(LogManager.getLogger(DateFormatters.class));
-    // pkg private for tests
-    static final Map<String, String> DEPRECATED_SHORT_TIMEZONES;
-    static {
-        Map<String, String> tzs = new HashMap<>();
-        tzs.put("EST", "-05:00"); // eastern time without daylight savings
-        tzs.put("HST", "-10:00");
-        tzs.put("MST", "-07:00");
-        tzs.put("ROC", "Asia/Taipei");
-        tzs.put("Eire", "Europe/London");
-        DEPRECATED_SHORT_TIMEZONES = Collections.unmodifiableMap(tzs);
-    }
-
-    public static ZoneId dateTimeZoneToZoneId(DateTimeZone timeZone) {
-        if (timeZone == null) {
-            return null;
-        }
-
-        String deprecatedId = DEPRECATED_SHORT_TIMEZONES.get(timeZone.getID());
-        if (deprecatedId != null) {
-            DEPRECATION_LOGGER.deprecatedAndMaybeLog("timezone",
-                "Use of short timezone id " + timeZone.getID() + " is deprecated. Use " + deprecatedId + " instead");
-            return ZoneId.of(deprecatedId);
-        }
-        return ZoneId.of(timeZone.getID());
     }
 }
